// web-application/src/types-openapi/api/models/CreatePayrollRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreatePayrollRequest
 */
export interface CreatePayrollRequest {
    /**
     * 
     * @type {string}
     * @memberof CreatePayrollRequest
     */
    name: string | null;
    /**
     * 
     * @type {Date}
     * @memberof CreatePayrollRequest
     */
    startDate: Date;
    /**
     * 
     * @type {Date}
     * @memberof CreatePayrollRequest
     */
    endDate: Date;
    /**
     * 
     * @type {string}
     * @memberof CreatePayrollRequest
     */
    note?: string | null;
}

/**
 * Check if a given object implements the CreatePayrollRequest interface.
 */
export function instanceOfCreatePayrollRequest(value: object): value is CreatePayrollRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('startDate' in value) || value['startDate'] === undefined) return false;
    if (!('endDate' in value) || value['endDate'] === undefined) return false;
    return true;
}

export function CreatePayrollRequestFromJSON(json: any): CreatePayrollRequest {
    return CreatePayrollRequestFromJSONTyped(json, false);
}

export function CreatePayrollRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreatePayrollRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'startDate': (new Date(json['startDate'])),
        'endDate': (new Date(json['endDate'])),
        'note': json['note'] == null ? undefined : json['note'],
    };
}

export function CreatePayrollRequestToJSON(json: any): CreatePayrollRequest {
    return CreatePayrollRequestToJSONTyped(json, false);
}

export function CreatePayrollRequestToJSONTyped(value?: CreatePayrollRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'startDate': ((value['startDate']).toISOString().substring(0,10)),
        'endDate': ((value['endDate']).toISOString().substring(0,10)),
        'note': value['note'],
    };
}